+++ src/main/java/com/shinesolutions/aemorchestrator/exception/MessageHandlerNotFoundException.java
@@ -1,10 +1,10 @@
 package com.shinesolutions.aemorchestrator.exception;
 
-public class MessageHandlerNotFound extends Exception {
+public class MessageHandlerNotFoundException extends Exception {
 
     private static final long serialVersionUID = 1L;
 
-    public MessageHandlerNotFound(String messageSubject) {
+    public MessageHandlerNotFoundException(String messageSubject) {
         super("No message handler found for message with subject: " + messageSubject);
     }